app/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dataclasses-json==0.5.14
decorator==5.1.1
Flask==2.3.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.251
langsmith==0.0.18
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.0
pycparser==2.21
pydantic==1.10.12
pypdf==3.14.0
pypdfium2==4.18.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
validators==0.20.0
weaviate-client==3.22.1
Werkzeug==2.3.6
yarl==1.9.2
